Sprint 2

+++ src/main/java/manske/locadora/manskemathes/controller/DiretorControler.java
@@ -3,8 +3,12 @@ package manske.locadora.manskemathes.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+import manske.locadora.manskemathes.model.Classe;
 import manske.locadora.manskemathes.model.Diretor;
 import manske.locadora.manskemathes.repository.DiretorRepository;
+import manske.locadora.manskemathes.repository.FilmeRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -15,8 +19,12 @@ public class DiretorControler {
 
     private final DiretorRepository diretorRepository;
 
-    public DiretorControler(DiretorRepository diretorRepository) {
+    private final FilmeRepository filmeRepository;
+
+
+    public DiretorControler(DiretorRepository diretorRepository, FilmeRepository filmeRepository) {
         this.diretorRepository = diretorRepository;
+        this.filmeRepository = filmeRepository;
     }
 
     @GetMapping
@@ -36,10 +44,24 @@ public class DiretorControler {
         return diretorRepository.save(diretor);
     }
 
+    private boolean validaExcluir(Diretor diretor) {
+        return filmeRepository.existsByDiretor(diretor);
+    }
+
     @DeleteMapping("/{id}")
     @ResponseStatus(code = HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
-        diretorRepository.deleteById(id);
+       Optional<Diretor> diretorEncontrado = diretorRepository.findById(id);
+
+        if (diretorEncontrado.isPresent()) {
+            if (validaExcluir(diretorEncontrado.get())) {
+               throw new ResponseStatusException(
+                HttpStatus.BAD_REQUEST, "Diretor tem relação com outra"
+            );
+            } else {
+                diretorRepository.deleteById(id);;
+            }
+        }
     }
 
     @PutMapping("/{id}")
