+++ src/main/java/manske/locadora/manskemathes/controller/LocadoraControler.java
@@ -11,9 +11,13 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
+import manske.locadora.manskemathes.model.Classe;
 import manske.locadora.manskemathes.model.Filmes;
 import manske.locadora.manskemathes.repository.FilmeRepository;
+import manske.locadora.manskemathes.repository.ItemRepository;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -23,9 +27,11 @@ import java.util.Optional;
 public class LocadoraControler {
 
     private final FilmeRepository filmeRepository;
+    private final ItemRepository itemRepository;
 
-    public LocadoraControler(FilmeRepository filmeRepository) {
+    public LocadoraControler(FilmeRepository filmeRepository, ItemRepository itemRepository) {
         this.filmeRepository = filmeRepository;
+        this.itemRepository = itemRepository;
     }
 
     @GetMapping
@@ -34,8 +40,8 @@ public class LocadoraControler {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity <Filmes> encontrarFilme (@PathVariable Long id){
-        Optional <Filmes> filmeEncontrado = filmeRepository.findById(id);
+    public ResponseEntity<Filmes> encontrarFilme(@PathVariable Long id) {
+        Optional<Filmes> filmeEncontrado = filmeRepository.findById(id);
         return filmeEncontrado.map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
@@ -45,10 +51,24 @@ public class LocadoraControler {
         return filmeRepository.save(movie);
     }
 
+    private boolean validaExcluir(Filmes filme) {
+        return itemRepository.existsByFilme(filme);
+    }
+
     @DeleteMapping("/{id}")
     @ResponseStatus(code = HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
-        filmeRepository.deleteById(id);
+        Optional<Filmes> filmeEncontrado = filmeRepository.findById(id);
+
+        if (filmeEncontrado.isPresent()) {
+            if (validaExcluir(filmeEncontrado.get())) {
+                throw new ResponseStatusException(
+                        HttpStatus.BAD_REQUEST, "Classe tem relação com outra");
+            } else {
+                filmeRepository.deleteById(id);
+            }
+        }
+
     }
 
     @PutMapping("/{id}")
@@ -56,13 +76,16 @@ public class LocadoraControler {
     public void update(@PathVariable Long id, @RequestBody Filmes movie) {
         Optional<Filmes> filmeEncontrado = filmeRepository.findById(id);
 
-        if (filmeEncontrado.isPresent()){
+        if (filmeEncontrado.isPresent()) {
             Filmes filme = filmeEncontrado.get();
             filme.setNome(movie.getNome());
             filme.setClasse(movie.getClasse());
             filme.setAno(movie.getAno());
             filme.setSinopse(movie.getSinopse());
             filme.setCategoria(movie.getCategoria());
+            filme.setDiretor(movie.getDiretor());
+            filme.setAtor(movie.getAtor());
+            // filme.setAtor(movie.getAtor());
 
             filmeRepository.save(filme);
         }
